// src/components/Forms/__generated__/MemberFieldlet_MemberInfo.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: MemberFieldlet_MemberInfo
// ====================================================

export interface MemberFieldlet_MemberInfo {
  __typename: "User";
  itemId: number;
  fname: string;
  sname: string;
  localAlias: string | null;
  eduroam: string | null;
}